Agrosis_API-dev/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.3.0
audioop-lts==0.2.1
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
distlib==0.3.8
Django==5.1
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-yasg==1.21.8
email_validator==2.2.0
Faker==37.1.0
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.15.4
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
mysqlclient==2.2.7
numpy==2.2.4
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.2.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.20
pyttsx3==2.98
pytz==2024.2
pywin32==308
PyYAML==6.0.2
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
service-identity==24.2.0
setuptools==76.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.1
sqlparse==0.5.1
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.0
toposort==1.10
Twisted==24.11.0
txaio==23.1.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.26.3
watchfiles==1.0.4
websockets==15.0.1
zope.interface==7.2
